+++ DarmuhsTerminalCommands/Compatibility/OpenBodyCamsCompatibility.cs
@@ -1,7 +1,20 @@
-﻿namespace TerminalStuff
+﻿using System.Runtime.CompilerServices;
+
+namespace TerminalStuff
 {
     internal class OpenBodyCamsCompatibility
     {
-        //moved to OpenLib
+        [MethodImpl(MethodImplOptions.NoInlining)]
+        internal static bool IsCreatingCommands()
+        {
+            if (!Plugin.instance.OpenBodyCamsMod)
+                return false;
+
+            if (OpenBodyCams.Plugin.TerminalPiPBodyCamEnabled.Value)
+                return true;
+
+            return false;
+
+        }
     }
 }